Using doulbe instead.

+++ terminus_est/src/mcts/base/IHeuristic.java
@@ -4,5 +4,5 @@ package mcts.base;
  * Calculates the "value" of a state.
  */
 public interface IHeuristic {
-    float Calculate(NodeMCTS state);
+    double Calculate(NodeMCTS state);
 }
